+++ lib/aspera/rest.rb
@@ -37,7 +37,9 @@ module Aspera
       user_agent:              'Ruby',
       download_partial_suffix: '.http_partial',
       # a lambda which takes the Net::HTTP as arg, use this to change parameters
-      session_cb:              nil
+      session_cb:              nil,
+      proxy_user:              nil,
+      proxy_pass:              nil
     }
 
     class << self
@@ -75,6 +77,21 @@ module Aspera
         end
         return uri
       end
+
+      def start_http_session(base_url)
+        uri = build_uri(base_url)
+        # this honors http_proxy env var
+        http_session = Net::HTTP.new(uri.host, uri.port)
+        http_session.proxy_user=proxy_user
+        http_session.proxy_pass=proxy_pass
+        http_session.use_ssl = uri.scheme.eql?('https')
+        http_session.set_debug_output($stdout) if debug
+        # set http options in callback, such as timeout and cert. verification
+        session_cb&.call(http_session)
+        # manually start session for keep alive (if supported by server, else, session is closed every time)
+        http_session.start
+        return http_session
+      end
     end
 
     private
@@ -82,15 +99,7 @@ module Aspera
     # create and start keep alive connection on demand
     def http_session
       if @http_session.nil?
-        uri = self.class.build_uri(@params[:base_url])
-        # this honors http_proxy env var
-        @http_session = Net::HTTP.new(uri.host, uri.port)
-        @http_session.use_ssl = uri.scheme.eql?('https')
-        @http_session.set_debug_output($stdout) if self.class.debug
-        # set http options in callback, such as timeout and cert. verification
-        self.class.session_cb&.call(@http_session)
-        # manually start session for keep alive (if supported by server, else, session is closed every time)
-        @http_session.start
+        @http_session=self.class.start_http_session(@params[:base_url])
       end
       return @http_session
     end